add init imports and create knackpy.py
